refactor: 일반 로그인/회원가입 archive 디렉토리로 이동
